// admin-workbench/lib/admin/types.ts
export type AdminUserItem = {
  id: number;
  role: "customer" | "vendor";
  name: string | null;
  email: string | null;
  phone_number: string | null;
  is_active: boolean;
  created_at: string;
};